Converted MOTM dashlet to "Mod Spotlights" dashlet
Dashlet needed to be less "monthly" (the tied workflow on Vortex-Backend
has been changed too)
